terminate test
