+++ integration_tests/src/tests/mod.rs
@@ -26,7 +26,8 @@ mod publish_deals_test;
 pub use publish_deals_test::*;
 mod replica_update_test;
 pub use replica_update_test::*;
-// mod terminate_test;
+mod terminate_test;
+pub use terminate_test::*;
 // mod verified_claim_test;
 // mod verifreg_remove_datacap_test;
 mod withdraw_balance_test;
